feat(ipc): extract ElectronAPI to electron-types and tighten createIpcListener
- Move ElectronAPI interface + Window.electronAPI global declaration into
  packages/shared/src/electron-types.ts per #62 spec; re-export from
  packages/shared/src/types.ts so existing @chamber/shared/types imports
  continue to resolve unchanged.
- Constrain createIpcListener<T>(ipcRenderer, channel, callback) to require
  IpcChannel from #61 instead of a free string. Misspelled or unknown
  channels now fail at compile time, making the IPC channel constants
  a real contract rather than decoration. Addresses Uncle Bob's deferred
  ask from PR #234 / #61.
- Migrate createIpcListener.test.ts to consume IPC.* constants and add
  @ts-expect-error coverage for both unknown literals and free string.
- New electron-types.test.ts pins the surface and the back-compat path.
- v0.47.2 patch bump + matching CHANGELOG entry.

Closes #62
